feat: renamed package to better reflect purpose
