+++ terraform/testdata/extracttest/main.tf
